Update patch set 1
Patch Set 1: Code-Review-1

(6 comments)

To save and restore the error code ${?} during 'administrivia' so that all tests can propagate to a wired-or error code result.

The administrivia was to ensure that once the test completed, that we did not leave an invalid value in persist.sys.boot.reason.

However, upon second glance, we do not _need_ a debug build to fix this up, we can perform a reboot instead with a 'healthy' value. I think I will restructure this test to be more resilient to "user" builds.

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Presubmit-Ready=0