Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > That's not possible -- there is not a /vendor/libdrm.so without BOARD_VNDK_VERSION.
> 
> This is because libdrm is vendor_available:true, right? I think we can change it back to vendor:true and define another libdrm (i.e. libdrm_platform) for libminui.

Yes. We've got vendor_available, we've got to be able to use it. Keeping old hacks alive is not a valid reason to not use the supported methods.

> Anyway, aosp_walleye will be bootable again when O-MR1 is published to AOSP and thus aosp/master is rebased. (since it can then be built with BOARD_VNDK_VERSION). So, I am okay with this change if this will be reverted back then.

Anything that was using this should be switched to BOARD_VNDK_VERSION (or doesn't need this if it's never going to switch)

> I do agree that the fake vndk-sp was nasty, but I think we have to keep it until we stop supporting Oreo version of vendor image.

This is master, we don't support the Oreo version of the vendor image.

Patch-set: 1
Reviewer: Dan Willemsen <1074628@85c56323-6fa9-3386-8a01-6480fb634889>