Update patch set 3
Patch Set 3: Code-Review-2

I talked to yabin and we want to get a stable simpleperf, so we want to hold off on this until next week.

In addition, to get the performance of the offline unwinder similar to the current code, I need to implement a system wide elf cache. That will probably come as part two though.

Patch-set: 3
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2