Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

agreed that it's a symptom of something else wrong, but as is I think we'll continue to boot and run while lmkd is spinning starting/stopping. if that happened on a user device for some reason, that would be catastrophic because lmk would never happen (in addition to lmkd spinning).

maybe drop the return 1 and somebody writes a CTS test to make sure that lmkd is actually pinned?

Patch-set: 1
Reviewer: Tim Murray <1017430@85c56323-6fa9-3386-8a01-6480fb634889>