Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> The program is locked into memory, so all pages are private dirty forever. If you don't want to take a patch of this form, we'll just write to dmesg instead.

Bionic folks, this is what it boils down to because lmkd needs to be pinned. Not sure if we have any other apps that need to be mlock'd right now, but if we ever do, they're going to hit the same problem re logging.

Patch-set: 2
Reviewer: Tim Murray <1017430@85c56323-6fa9-3386-8a01-6480fb634889>