Update patch set 2
Patch Set 2: Code-Review+2

> This patch isn't about C vs C++. We're conditionally compiling out some code backing a feature. That this feature is generated by C++ code is irrelevant.

I thought from the commit message that the point was not to pull in libc++.  I just figure that that's going to limit you in the future.  Unless the event map is actually causing a decent bit of the savings too.

In any case, I'm pretty okay with the patch.

Patch-set: 2
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2