Update patch set 2
Patch Set 2:

Hi John,

The recommendation in https://android-review.googlesource.com/c/kernel/common/+/292158 is out of date. Functionfs is a pseudo-filesystem with well known behavior that does not require buffering and as such AIO is a reasonable interface to interact with it. AIO also provides performance improvements that are not possible with other existing apis (read(), write()). You are right that android-recommended.cfg needs to be updated to remove the recommendation to turn AIO off.

Wrt the patch in particular, having io_* syscalls disabled in config is one reason that aio would fail, but it is also possible that the syscalls are enabled, but functionfs itself doesn't implement them (true in <3.18). This can't be determined programatically since it would require sending or receiving an actual packet, so we have the property. We will look into documenting it as well.

Patch-set: 2
Reviewer: Jerry Zhang <1113594@85c56323-6fa9-3386-8a01-6480fb634889>