Update patch set 2
Patch Set 2:

(6 comments)

Thanks Mark!

We've vacillated back and forth on this a few times.  Just to be clear:

1)  We talked about CarPowerManager using the libsuspend library in its own process space.  This creates the complexity of two instances of libsuspend running in separate processes, racing with each other.  Not ideal, and a bit messy.

2)  If we're able to put forceSuspend into Android Power Manager Service (aosp/582371), then CarPowerManager will call forceSuspend() which will go through Android Power Manager's process.  This would keep all power manager functionality in the same process and avoid the complexity above.

If we are able to do (2), then I would like to avoid (1) altogether.

In either case (whether we do (1) or (2)), I believe forceSuspend needs to perform the wakeup_count dance that's encoded in the thread.  So, even if we go with (1), we still need to do all that work.

If we take the KISS forceSuspend solution, it will take the logic in the thread and put it into a separate function, which cannot be shared with autosuspend (reading wakeup_count is a blocking call).  Hence, we still need to run the init() function, even for KISS forceSuspend.

Patch-set: 2
Reviewer: Steve Paik <1061103@85c56323-6fa9-3386-8a01-6480fb634889>