Update patch set 2
Patch Set 2:

(1 comment)

Part of the vacillation is a result of my prototype, where I wanted the native layer to be-all, end-all for the entire process. I have to admit there are many things better handled higher up in the chain. Until recently, I did not know what part of the stack you were working on, and where the sections were. I also perceived there is a standard cookbook of things we could do, and parts that are specific to the hardware.

Questions:
1) what can you do to turn off the sensorhub, and other unnecessary services/hardware to minimize power consumption?
2) what can we do to minimize wake sources to the subset, and who should control this?
3) What call, and at what level, can we make in preparation for disabling autosuspend
4) When do we want autosuspend enabled? (any system that supports force suspend, automagically support the lighter touch of autosuspend)

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>