Update patch set 2
Patch Set 2:

Wait actually.  Is there a separate .rc file next to the daemon that uses these directories?  My intention a while back when we split out the different .rc files was that even these commands would go into them.  In theory a system might not use that daemon nor need these directories in which case they wouldn't need to be created.

Unless it'll be used by multiple daemons in which case sure this place is fine.  But I've been hoping to stop using this as a general dumping ground.

Patch-set: 2
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>