Update patch set 8
Patch Set 8:

Fyi, this isn't really ready yet.  As mentioned here / on the bug, I'm going to drop the min/max and rename 'schema' to 'type'.  It'll be a much cleaner design and it doesn't seem like the min/max would be used in practice.

Patch-set: 8
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>