Update patch set 1
Patch Set 1:

Still needs some work, but posting now for comments.  This seems to be the general idea that I think must be done.  

Alternatives are actually breaking property service out of init or running it in a different thread, which makes less and less sense as time goes on.

The function pointer is because not just do_setprop potentially sets a property, but any init function can (and actually can set multiple properties), so we need to make sure that all of these properties are set.

Patch-set: 1
Reviewer: Jaekyun Seok <1019198@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>