Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I think I've answered to all questions. Is there any concern left?
> 
> Yes - the concern about punishing good actors has not been addressed.

Hmm, guess that question has no answer because that's not about true/false, but about right/wrong, which is rather subjective.

One last thing I would like to comment is that there currently is no good actor on this, because this vendor/system separation is a new requirement since Treble and independent system OTA is still possible only in a limited circumstance. Before Treble, they had access to the libs in the system partition unconditionally and thus they didn't have to be cautious on the existence of the lib.

But, yes, this blocks potential good actors. However, I think, in that case, the app should be part of the system image. This vendor -> system dependency is strictly disallowed in every parts of Treble (build system, sepolicy, sysprops, ...) I don't think the potential good actors can be a good excuse to break the established rule.

Patch-set: 1
Reviewer: Jiyong Park <1132673@85c56323-6fa9-3386-8a01-6480fb634889>