Update patch set 1
Patch Set 1:

This is not going to be submitted.  I'm sorry that you're having trouble, but this is canonical already; init commands are actions and tend to be verbs, hence 'setrlimit'.  Service options are 'options' and tend to be nouns, 'class', 'namespace', 'user', etc (setenv being an unfortunate exception).

What you probably really want is b/36970783, that's filed to check init scripts for correctness at build time.  It's on my list but it's been blocked for various reasons.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>