Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> how have you evaluated this on an actual system?

Tim, I think you were on all these reviews on internal master, weren't you? Didn't you have discussions with Suren on this and agreed that LMKD should be turned on for the 2018 devices, but not the other Pixel/Nexus. The Go team is already using a version of this.

Patch-set: 1
Reviewer: Todd Kjos <1069543@85c56323-6fa9-3386-8a01-6480fb634889>