Update patch set 5
Patch Set 5:

The return value of prctl is 'int' type as below.
- int prctl(int option, unsigned long arg2, unsigned long arg3,
                 unsigned long arg4, unsigned long arg5);
Also, you define securebits is as 'unsigned long', but it is compared to negative value. It doesn't make sense.
Now we met fault issue on 'if (securebits == -1UL)', so it could not boot-up. But, it's okay if we change unsigned log to int. 
How do you think about it?

Patch-set: 5
CC: Randy Jeong <1083259@85c56323-6fa9-3386-8a01-6480fb634889>