Update patch set 5
Patch Set 5:

Do you have the unsigned integer sanitizer on?  You'll blow up if it is.  We removed it with https://android-review.googlesource.com/c/platform/system/core/+/445032.

Changing to ~0UL should let us leave on the unsigned integer sanitizer, but as the above commit message states, this isn't the first time we've blown up unnecessarily due to it.

Patch-set: 5