Update patch set 2
Patch Set 2: Code-Review+1

(1 comment)

Since this code seems to be WAI, no pressure to address my comment. :)

If the analyzer complains, we can just //NOLINT this, or something. Part of me is hoping it won't, since we're sending the pointer to LOG(), but we'll see.

Patch-set: 2
Reviewer: George Burgess IV <1096946@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1