Update patch set 1
Patch Set 1:

"When DbC is initialized and enabled, it will present a debug device through the debug port (normally the first USB3 super-speed port)."

Does the debug port work if the Android is connected to a superspeed hub? Which port would it choose as the debug port?

Patch-set: 1
Reviewer: Jerry Zhang <1113594@85c56323-6fa9-3386-8a01-6480fb634889>