Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 1:
> > > 
> > > "When DbC is initialized and enabled, it will present a debug device through the debug port (normally the first USB3 super-speed port)."
> > > 
> > > Does the debug port work if the Android is connected to a superspeed hub? Which port would it choose as the debug port?
> > 
> > We can add hub to the Debug host. A more detail information of the topology is in xhci specification 1.1 in section 7.6.1.
> 
> Our desired use case is to add a hub to the debug target. I see in the spec that this is currently specifically disallowed.
> 
> "A Debug Target may only connect to a Debug Host through a Root Hub port of the target."
> 
> Is this a technical limitation of a usb hub? One of our desired use cases is to debug a Android in host mode while it is connected to a device (such as usb headphones). Since pretty much all android phones have only one usb port, a hub is needed.

Please note that this solution is for "host platforms" with xHCI ports. A typical scenario is to connect two laptops back to back.

When we put a port into DbC mode the port is re-purposed to USB device mode and hence we may not be able to connect a USB hub.

So in a mobile scenario where we have a single port we may not be able use this method, But we could adopt other methods by developing split cards assuming port is USB-C

Patch-set: 2