Update patch set 3
Patch Set 3: Code-Review+1

(2 comments)

Hmm, for the record this has come up before, see CL:454276 where made avbctl require the user to use --force if the device is locked.

FWIW, I'm generally not a big fan of making software protect itself from users.. a least not without giving the user an out (e.g. the --force option) ... unfortunately it's not very easy to add "--force" to adb since it's client server and the client typically is an older version. But let's at least mention that 'avbctl --force disable-verity' is an option?

I'm not in the OWNERS file for adb so can't really give you a +2 ... but here's a +1 for the AVB bits.

Patch-set: 3
Reviewer: David Zeuthen <1076971@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1