Update patch set 1
Patch Set 1: Code-Review+2

Is there not a better place for this file too though?  Why not put all of this in bionic/linker/etc or something closer to the consumer itself?  I'm not a huge fan of having a generic source folder for what goes to /etc, especially when these files have a specific consumer, whose source they can be moved close to.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2