Update patch set 1
Patch Set 1:

Sounds good to me, especially moving hosts/.  We could deprecate the /etc directory here if we move these all.

In general, the more we remove from rootdir the better in my opinion.  Given that we don't really have a ramdisk anymore, there's virtually nothing that belongs in /.  I'd be happy to remove the /etc directory and state that we will not add any more files to this directory.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>