Update patch set 1
Patch Set 1: Verified+1 Presubmit-Ready+1

As evident on a 4.9 kernel (hikey). access(INKERNEL_MINFREE_PATH, W_OK) still returns 0 if the node existed before this change, so this does not fix a bug in the test, but there is a possibility that there is a kernel bug (?) that causes the selinux audit noise. Since the in kernel lowmemorykiller is being deprecated, let's not care about that problem ...

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Presubmit-Ready=+1