Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I think I see how this is needed to make http://ag/3801480 continue to work. Can you remind me, though, of what we're committing to, by having this property whitelisted? (I want to make sure we can uphold "our end of the bargain".)

Once a property is whitelisted, it should be kept as stable as possible. Renaming it wouldn't be allowed, and deprecating it would be processed carefully with some time window not to break ABI between system and vendor partitions.

Patch-set: 1
Reviewer: Jaekyun Seok <1019198@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0