Update patch set 3
Patch Set 3:

(2 comments)

> Patch Set 3:
> 
> > Patch Set 1:
> > 
> > What does this help with?  The bug reference healthd having an issue logging to dmsg, but the comment here says that init is the only user of KernelLogger?
> 
> Because, init is the only user _now_, it speaks volumes that access to KernelLogger is currently broken as compared to libcutils klog facility, it needs to do this so that others that are not init have access.

This is broken in that if the process doesn't have access to /dev/kmsg that it won't be able to write to the kernel log?  Init has full access, so it doesn't hit it, but other daemons may not have the permissions?  Particularly the one mentioned in the bug?  It may help to make that clear in the commit message.

> 
> > 
> > Is there an associated change that passes /dev/kmsg to any services to populate the environment variable?
> 
> rc command:
> 
>     file w /dev/kmsg

Right, is there a CL for adding this to any other daemons?

Patch-set: 3
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>