Update patch set 1
Patch Set 1:

is this widely used already? it's a terrible API that keeps growing extra random arguments, and still doesn't cover a lot of the would-be use cases.

there are plans for a new API in Q if the point of this is to offer new functionality rather than to keep old code building...

Patch-set: 1
CC: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>