Update patch set 1
Patch Set 1: Code-Review+2

This is IMHO a more than adequate and simple fix for the associated bug for upcoming release.

[TL;DR]

Has it been confirmed that this also eliminates storaged trying to use the reserved region of the filesystem (and thus get an sepolicy hit and subsequent crash)?

More work could follow for later releases. For instance the daemon could be run as non-root and still accomplish it goals? (Some /proc nodes do not support DAC/MAC overrides, requiring a root daemon, attributes that come and go can also create problems with access if not solved by readproc gid)

Is there a storaged adjustment to cope with write failures? Currently I suspect it writes it, does not complain, or does and crashes, or maybe when reading the values back crashes? Neither is good because the in-memory data is accumulated and can be lost. It should detect a write, or partial read, failures and keep on going accumulating information with the hope that it will eventually land the data (in the non-reserved blocks)? Perhaps relaxing its schedule when ENOSPC is returned from any write operation, deprioritizing itself from writing anything so other more critical elements can have their way with the storage.

Patch-set: 1
Reviewer: Jeffrey Vander Stoep <1054468@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2