Update patch set 1
Patch Set 1:

I have series pending for review based on the fdo drm_hwcomposer but I haven't finished testing it yet. This problem I have tested and worked around differently. For simplicity, I uploaded it against AOSP, but feel free to cherry-pick this over to your repository.

The libsync library is just an ioctl wrapper so I think it's OK for you guys to keep using the obsolete features in a local build of it, as you have already acknowledged that we should remove this code asap and move to in/out fencing.

https://android-review.googlesource.com/#/c/platform/external/drm_hwcomposer/+/673592

Patch-set: 1
Reviewer: Alistair Strachan <1293738@85c56323-6fa9-3386-8a01-6480fb634889>