Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> So we may not need this CL if we decide to use the /misc partition....

I understand that the /misc partition is used to bypass the late arrival of bootstat (post mount of /data) and that this is early in the development cycle. It would be 'nice' if we found a secure means to update the reboot reason(s) to aid triage. Ideally we want this to come via the bootloader into ro.boot.bootreason so you do not have to manhandle storage.

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>