Update patch set 1
Patch Set 1:

(5 comments)

Draft reviews really makes reviews confusing in the new interface. The only question I wanted answered was if the idea of this code (currently 300 lines, could shrink more) was acceptable to be maintained in init. I assume because you gave it a deeper review (than expected) that you just answered yes to the question?

I can refine it _much_ tighter if given the go-ahead, I did not want to waste too much time on the first pass. I will refactor to move /dev/keychord driver portion as an afterthought add-on as requested and turn this into a real (public) review tomorrow, and actually put some time into it rather than just prototyping and performing feasibility/performance testing.

I can not submit it until after the std::function code has gone in.

This code, as is with a CP of std::function, has successfully triggered a bugreport with either /dev/keychord or /dev/input when available. I have confirmed that even though EVIOSCMASK does not work, that the interruption rate is still not high enough to worry about, so we are willing to _drop_ /dev/keychord functionality.

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0