Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > > Patch Set 9:
> > > 
> > > (1 comment)
> > > 
> > > I think this looks okay, but I'm planning to do a release of platform-tools soon, so I'm going to sit on this for a bit (snapping the branch to aosp/master Soon, but I want to get a few other patches in first to not have to go through cherry-pick hell)
> > 
> > aw, i was hoping to get this in before the release so that the cts harness wouldn't have to use custom-built versions of adb. oh well, if this gets in, at least we'll be able to use the stock version next year.
> > 
> > thanks for the reviews!
> 
> is the CTS harness _already_ doing that (and this patch lets you remove it), or is this a yet-to-be-written workaround?

The official one does have any sort of workaround for this. The one Chrome OS uses is a thin layer around the official one, and one of the mods is a hacked-up Python script that continuously polls the state of the connection used to work for N but started failing for P. Instead of trying to pile more hacks on top of our previous hacks (and to slowly try to remove the layers around the official harness), I tried to fix the root cause.

Patch-set: 9
Reviewer: Luis Hector Chavez <1066743@85c56323-6fa9-3386-8a01-6480fb634889>