Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > (1 comment)
> > 
> > Under what conditions will this be used? As long as you are willing to allow a crash, then this will be fine, but I've never found a case where you can guarantee no crashes.
> 
> To record a call stack of the current thread for an invocation of open/close. Is it actually possible for this to crash if the registers/stack are in a good state and the unwind information isn't garbage?

There is a high likelihood that if you are unwinding on arm32 in the main thread, it will crash. Without modifying a lot of code, I couldn't get the unwind to terminate properly and since it was only the main thread, I left it alone. Also, unwinding on host would likely also crash (same problem, the libc libraries on newer systems don't have an exact terminating condition).

I think that in most other cases, it should be okay, but I still can't guarantee no crashes. I think the only other way you would see a crash if the unwind terminates normally is if there is some kind of stack corruption or memory corruption.

Patch-set: 5