Update patch set 10
Patch Set 10:

Looks good to me.  We'll need to figure out how to communicate to others that the 'boot_devices' DT entry needs to be used for cppreopts, etc.

I'm not thrilled to say this, but it's why I wondered if we do need to parse the fstab in DT to make a guess at what the 'boot_devices' are when that parameter isn't defined.  Especially given Treble, we'd have to assume that a system image with this change will be used with a boot image without boot_devices present.

Patch-set: 10
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>