Update patch set 2
Patch Set 2:

(5 comments)

Some thoughts inline.

We should probably turn keychords.cpp into an actual class that can be instantiated too in this CL.  If I'm imagining this right, it should be able to be local to init's main() and not needing a singleton.

Patch-set: 2
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>