Update patch set 3
Patch Set 3:

FileMap was only introduced to libziparchive because that code needed to run on Windows. 

At the time, we debated spliting libutils out (or even copy pasta of the few lines of windows code that we need). I regret not doing it because this dependency has caused a looot of trouble since it was introduced.

Patch-set: 3
CC: Narayan Kamath <1014443@85c56323-6fa9-3386-8a01-6480fb634889>