Update patch set 3
Patch Set 3:

For the addition of a timeout or debounce time (a solitary trailing negative keycode turns into in ms) in https://android-review.googlesource.com/c/platform/system/core/+/689583 the order is important for simplicity (first element is always the timeout), but I can certainly have it scan for the negative value since that is at construction time inside Keychord class and would add just 2 lines ...

Since the typical set size is 3 elements, I believe std:unordered_set buckets are memory inefficient as compared to a b-tree list?

NB: the debounce CL is currently exploratory/optional to demonstrate a facility that would allow a 3s long press of the power button to trigger a kernel panic.

I do not fully understand the documented proviso for equal comparison unique to std::unordered_set:

"The (==) behavior is also undefined if Hash and KeyEqual do (until C++20)KeyEqual does (since C++20) not have the same behavior on lhs and rhs or if operator== for Key is not a refinement of the partition into equivalent-key groups introduced by KeyEqual (that is, if two elements that compare equal using operator== fall into different partitions)"

I expect this is a non-issue b/c <int> is simple Key type and likely does not have any of these issues.

Patch-set: 3
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>