Update patch set 3
Patch Set 3:

(2 comments)

Agreed at the maintenance aspect.  I really just figure if we're doing equality comparisons, including when checking for duplicate entries during registration that a set would be easier.  These equality checks happen extremely infrequently and are ~3 entries each, so I don't think there'll be any real world performance impact on any decisions here.

That said, if this does get less maintainable with a set, then it's probably find to just continue to use a vector.  Or even a class that contains a vector and provides the necessary equality functions.

> first element is always the timeout

I wouldn't do this.  Once there is a timeout, I'd unwrap it into a data structure with an explicit 'timeout' element and a set/vector of the keycodes.

> Since the typical set size is 3 elements, I believe std:unordered_set buckets are memory inefficient as compared to a b-tree list?

Both are negligible.

> I do not fully understand the documented proviso for equal comparison unique to std::unordered_set:

I'm not sure myself.  That's why I was figuring set, since it's clearly equality comparable.  unordered_set seems like it'd need to play the same tricks that vector would need to be comparable.  At least it'd be hidden from us.

Patch-set: 3
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>