Update patch set 11
Patch Set 11:

NB: still -2 b/c it is an alternate approach that is not _necessary_ but may be helpful. The problem I have with the current solution for b/79553608 is it requires changes in bootloaders, interactions with a small list of experts, and in vendor systems; whereas this approach can be used device independent. The disadvantage of this approach is it assumes the kernel has not livelock'd in the input event system drivers and/or init has not livelock'd in user space; whereas long press power is an NMI picked up by the bsp and thus guaranteed.

Patch-set: 11
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>