Update patch set 14
Patch Set 14:

ditto, A feature in search of user ;-}.

NB: I explored this b/c bdc@ said 'interesting idea' as part of the llkd work ... The only other possible users (ClockWork) abandoned the debounce request as a limited-button bugreport trigger. Despite caution, I expect this to be >98% functional for this outlined example. This was an easy add-on that was part of a more serious patch set and could remain as an accessible bit-rotting Gerrit example if nothing else...

Patch-set: 14
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0