Update patch set 1
Patch Set 1: Code-Review-1

According to the spec, we already broke the parsing, because all options destined to go to init for interpretation are supposed to be init.<key>= format and are in-fact passed to argv[] ;-/

For a test, I was hoping that it made sense to move this more generic quote handling into libbase where a gTest made sense. If strictly kernel parsing as you are pushing, then doesn't make sense for libbase. Maybe init could use whatever we formulate here from libfs_mgr for one-stop-shop?

The kernel itself looks for <key>= (matching '-' to '_'), we should do the same then.

There are no gTests in fs_mgr (BAD BAD BAD) so about time to start ... but I did not want to cross that bridge until no hope to move into libbase. Will spin in a test.

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Presubmit-Ready=0