Update patch set 1
Patch Set 1:

Might as well take it.  As with the other CL, it would be nice to put a tiny static library for cmdline parsing in a place where init and fs_mgr can access.  We don't want these two things to be parsing the same command line in different ways.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>