Update patch set 2
Patch Set 2:

In that case, then maybe let's just drop this?  We've survived for this long without it and I don't see any clear use case where it's necessary.  We'll be a little bit more restrictive than the kernel, but that's fine.

Patch-set: 2
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>