Update patch set 1
Patch Set 1: Verified+1 Autosubmit+1 Presubmit-Ready+1 Code-Review+2

See aosp/679405 for previous approvals. This is the same change, so not bothering to re-review. I cherry-picked this to internal master and did (a) a checkbuild in the target that broke on the previous attempt, and (b) a TreeHugger presubmit test. Both passed, so submitting this again.

Patch-set: 1
Reviewer: Jesse Hall <1010672@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Code-Review=+2
Label: Autosubmit=+1
Label: Presubmit-Ready=+1