Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> thanks, rebased to pick up the newfs_msdos change so need a new +2
> 
> I assume that installs only to system? If so, newfs_msdos was previously installed on vendor and recovery too, was that removal intentional?

+jsharkey

thanks for checking, but, yes, that was deliberate. historically everything was in toolbox so we didn't have much choice. when we pulled newfs_msdos out, jsharkey felt that there's no use for newfs_msdos on /vendor so we shouldn't make it available without good reason.

(of course, we also have the opposite situation where things move into toybox and thus become available on /vendor "just because". if that ever becomes a problem, we could potentially build toybox with two different sets of generated files. making that work in the Android build system would probably be pretty unpleasant, though, so we'll cross that bridge if we ever come to it.)

Patch-set: 2
Reviewer: Jeff Sharkey <1013508@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>