Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+2
> 
> (3 comments)
> 
> > Do we want to remain functional if dm-zero is disabled? It means sparse regions will be backed by (precious) storage, but at least we have a volume to work with.
> 
> I'd say abort.  Dm-zero isn't enabled by default, but it's a pretty trivial thing to enable, right?  If you want to use it, better to mandate it than have a less desirable fall back.

dm-zero seems beneficial and fairly innocuous, so I'm in favor of requiring it. The original comment was more recognizing the current state of things, and I didn't want to submit a test that is guaranteed to fail.

If it's not present, the LoadTableAndActivate call will fail, and presumably the caller should abort. (Or we can validate that each target type appears in the available target type list).

Patch-set: 2
Reviewer: David Anderson <1290458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0